changing assumptions on execution

+++ service.py
@@ -1,29 +1,25 @@
 import zmq
 import time
 import json
-import os, sys
+import os, sys, signal
 
-class Repo(object):
+class Repo:
 	"""docstring for Repo"""
 	def __init__(self, path):
 		super(Repo, self).__init__()
 		self.path = path
-		self.mapping = {}
 
-	def refresh(self):
-		self.mapping = {}
+	def get_mapping(self):
+		mapping = {}
 		for current_path, dirs, files in os.walk(self.path):
 			for file in files:
 				source_path = os.path.join(current_path, file)
 				relative_path = os.path.relpath(source_path, source_root)
-				self.mapping[relative_path] = source_path
+				mapping[relative_path] = source_path
+		return mapping
 
-	def get_mapping(self):
-		if not self.mapping:
-			self.refresh()
-		return self.mapping
-
-class Lens(object):
+class Lens:
+	#TODO: track dirty edits
 	"""docstring for Lens"""
 	def __init__(self, out_dir, repos):
 		super(Lens, self).__init__()
@@ -54,12 +50,29 @@ class Lens(object):
 			for rel_path, source_path in dict.items():
 				dest_path = os.path.join(self.out_dir, rel_path)
 				os.makedirs(os.path.dirname(dest_path), exist_ok=True)
-				os.symlink(source_path, dest_path)
+				#take this out, eventually
+				if not os.path.isfile(dest_path):
+					os.symlink(source_path, dest_path)
+				else:
+					print("File already exists: %s" % dest_path)
 
 		write_symlinks(merged)
 
+class SvcState:
+	def __init__(self, cfg):
+		self.lenses = []
+		lensdefs = cfg['lenses']
+		for lens in lensdefs.keys():
+			repos = [Repo(repo) for repo in lensdefs[lens]]
+			self.lenses.append(Lens(lens, repos))
+
+
 def command_list(state, msg):
-	print("List command!")
+	print("Listening lenses and repos")
+	for lens in state.lenses:
+		print("Lens with output dir: %s" % lens.out_dir)
+		for repo in lens.repos:
+			print("  Repo source: %s" % repo.path)
 
 route =	\
 {
@@ -76,9 +89,15 @@ def get_socket(port):
 	socket.bind("tcp://*:%s" % port)
 	return socket
 
+def signal_handler(signal, frame):
+	print("Thank you, goodbye")
+	sys.exit(0)
+
 def main():
 	config = load_config()
 	socket = get_socket(config['port'])
+
+	state = SvcState(config)
 
 	while True:
 		msg_str = socket.recv_string()
@@ -86,9 +105,10 @@ def main():
 		msg = json.loads(msg_str)
 		print("Type: %s " % msg['type'])
 
-		route[msg['type']](config, msg)
+		route[msg['type']](state, msg)
 
 		time.sleep(1)
 
 if __name__ == '__main__':
+	signal.signal(signal.SIGINT, signal_handler)
 	main()
